refactor: improve code organization and use @appium/support utilities
- Move getMaxFrameLength to common utilities to avoid code duplication
- Replace custom toReadableSizeString with @appium/support util.toReadableSizeString
- Extract debugging logic into separate methods for better readability
- Fix indentation issues in LengthBasedSplitter
- Add JSDoc documentation for new debugging methods
