--- docs/README.md
@@ -1,15 +1,0 @@
-# Documentation
-
-This directory contains additional documentation for the `appium-ios-device` library.
-
-## Available Guides
-
-* [Debugging Large Frame Size Issues](DEBUG_FRAME_SIZE.md) - Comprehensive guide for troubleshooting "The frame is bigger than expected" errors with iOS devices
-
-## Contributing
-
-When adding new documentation:
-1. Use clear, descriptive titles
-2. Include practical examples
-3. Provide step-by-step instructions
-4. Update this README with links to new guides
