+++ lib/afc/index.js
@@ -9,22 +9,13 @@ import B from 'bluebird';
 import path from 'path';
 import _ from 'lodash';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 
 
 const AFC_SERVICE_NAME = 'com.apple.afc';
 const MAX_FRAME_SIZE = 1 * MB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 const NULL_DELIMITER_CODE = 0x00;
 const IGNORED_PATHS = ['.', '..'];
@@ -36,7 +27,7 @@ class AfcService extends BaseServiceSocket {
     this._splitter = new LengthBasedSplitter({
       readableStream: socketClient,
       littleEndian: true,
-      maxFrameLength: getMaxFrameLength(),
+      maxFrameLength: getMaxFrameLength(MAX_FRAME_SIZE),
       lengthFieldOffset: 8,
       lengthFieldLength: 8,
       lengthAdjustment: -8,
