+++ lib/house-arrest/index.js
@@ -5,22 +5,13 @@ import { KB } from '../constants';
 import { AfcService } from '../afc';
 import B from 'bluebird';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 
 
 const HOUSE_ARREST_SERVICE_NAME = 'com.apple.mobile.house_arrest';
 const MAX_FRAME_SIZE = 1 * KB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 class HouseArrestService extends BaseServiceSocket {
   constructor (socketClient) {
@@ -30,7 +21,7 @@ class HouseArrestService extends BaseServiceSocket {
     this._splitter = new LengthBasedSplitter({
       readableStream: socketClient,
       littleEndian: false,
-      maxFrameLength: getMaxFrameLength(),
+      maxFrameLength: getMaxFrameLength(MAX_FRAME_SIZE),
       lengthFieldOffset: 0,
       lengthFieldLength: 4,
       lengthAdjustment: 4,
