+++ lib/notification-proxy/index.js
@@ -4,22 +4,13 @@ import { LengthBasedSplitter } from '../util/transformer/length-based-splitter';
 import { KB } from '../constants';
 import _ from 'lodash';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 
 
 const NOTIFICATION_PROXY_SERVICE_NAME = 'com.apple.mobile.notification_proxy';
 const MAX_FRAME_SIZE = 16 * KB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 const RELAY_NOTIFICATION = 'RelayNotification';
 const PROXY_DEATH = 'ProxyDeath';
@@ -32,7 +23,7 @@ class NotificationProxyService extends BaseServiceSocket {
     this._splitter = new LengthBasedSplitter({
       readableStream: socketClient,
       littleEndian: false,
-      maxFrameLength: getMaxFrameLength(),
+      maxFrameLength: getMaxFrameLength(MAX_FRAME_SIZE),
       lengthFieldOffset: 0,
       lengthFieldLength: 4,
       lengthAdjustment: 4,
