+++ lib/plist-service/index.js
@@ -4,22 +4,13 @@ import { PlistServiceEncoder } from './transformer/plist-service-encoder';
 import { PlistServiceDecoder } from './transformer/plist-service-decoder';
 import { LengthBasedSplitter } from '../util/transformer/length-based-splitter';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 import { MB } from '../constants';
 
 
 const MAX_FRAME_SIZE = 1 * MB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 const CHECK_FREQ_MS = 50;
 
@@ -31,7 +22,7 @@ class PlistService extends BaseServiceSocket {
     this._splitter = new LengthBasedSplitter({
       readableStream: socketClient,
       littleEndian: false,
-      maxFrameLength: getMaxFrameLength(),
+      maxFrameLength: getMaxFrameLength(MAX_FRAME_SIZE),
       lengthFieldOffset: 0,
       lengthFieldLength: 4,
       lengthAdjustment: 4,
