+++ lib/usbmux/index.js
@@ -10,23 +10,14 @@ import path from 'path';
 import { PlistService } from '../plist-service';
 import { Lockdown, LOCKDOWN_PORT } from '../lockdown';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 import { MB } from '../constants';
 import { log } from '../logger';
 
 
 const MAX_FRAME_SIZE = 1 * MB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 const USBMUX_RESULT = {
   OK: 0,
@@ -124,7 +115,7 @@ class Usbmux extends BaseServiceSocket {
     this._splitter = new LengthBasedSplitter({
       readableStream: socketClient,
       littleEndian: true,
-      maxFrameLength: getMaxFrameLength(),
+      maxFrameLength: getMaxFrameLength(MAX_FRAME_SIZE),
       lengthFieldOffset: 0,
       lengthFieldLength: 4,
       lengthAdjustment: 0,
