+++ lib/util/transformer/length-based-splitter.js
@@ -1,6 +1,6 @@
 import Stream from 'stream';
 import { log } from '../../logger';
-import { toReadableSizeString } from '../../utilities';
+import { util } from '@appium/support';
 
 // Debug configuration
 const DEBUG_CONFIG = {
@@ -65,31 +65,13 @@ export class LengthBasedSplitter extends Stream.Transform {
       log.debug(`Frame size detected: ${messageLength} bytes, max allowed: ${this.maxFrameLength} bytes`);
     }
     
-          // Log large frames for investigation
-      if (DEBUG_CONFIG.enableLargeFrameLogging && messageLength > DEBUG_CONFIG.largeFrameThreshold) {
-        log.warn(`Large frame detected: ${toReadableSizeString(messageLength)}`);
-        log.warn(`Service: ${this.serviceName}, Max allowed: ${toReadableSizeString(this.maxFrameLength)}`);
-        
-        // Log memory usage if monitoring is enabled
-        if (DEBUG_CONFIG.enableMemoryMonitoring) {
-          const memUsage = process.memoryUsage();
-          log.warn(`Memory usage - RSS: ${toReadableSizeString(memUsage.rss)}, Heap: ${toReadableSizeString(memUsage.heapUsed)}`);
-        }
-      }
+    // Log large frames for investigation
+    if (DEBUG_CONFIG.enableLargeFrameLogging && messageLength > DEBUG_CONFIG.largeFrameThreshold) {
+      this._logLargeFrame(messageLength);
+    }
     
     if (messageLength > this.maxFrameLength) {
-      // Enhanced error message with debugging info
-      const errorMsg = `The frame is bigger than expected. Length: ${toReadableSizeString(messageLength)}, max: ${toReadableSizeString(this.maxFrameLength)}`;
-      log.error(errorMsg);
-      log.error(`Service: ${this.serviceName}, Frame offset: ${this.lengthFieldOffset}, Frame length: ${this.lengthFieldLength}`);
-      
-      // Suggest debugging steps
-      log.error('To debug this issue, set environment variables:');
-      log.error('  APPIUM_IOS_DEBUG_FRAME_SIZE=true');
-      log.error('  APPIUM_IOS_LOG_LARGE_FRAMES=true');
-      log.error('  APPIUM_IOS_MONITOR_MEMORY=true');
-      
-      throw new Error(errorMsg);
+      this._handleFrameSizeError(messageLength);
     }
 
     const completeMessageLength = messageLength + this.lengthAdjustment + this.lengthFieldOffset;
@@ -134,6 +116,39 @@ export class LengthBasedSplitter extends Stream.Transform {
 
   shutdown () {
     this.isShutdown = true;
+  }
+
+  /**
+   * Log information about large frames for debugging purposes
+   * @param {number} messageLength - The size of the detected frame
+   */
+  _logLargeFrame(messageLength) {
+    log.warn(`Large frame detected: ${util.toReadableSizeString(messageLength)}`);
+    log.warn(`Service: ${this.serviceName}, Max allowed: ${util.toReadableSizeString(this.maxFrameLength)}`);
+    
+    // Log memory usage if monitoring is enabled
+    if (DEBUG_CONFIG.enableMemoryMonitoring) {
+      const memUsage = process.memoryUsage();
+      log.warn(`Memory usage - RSS: ${util.toReadableSizeString(memUsage.rss)}, Heap: ${util.toReadableSizeString(memUsage.heapUsed)}`);
+    }
+  }
+
+  /**
+   * Handle frame size errors with enhanced debugging information
+   * @param {number} messageLength - The size of the frame that exceeded the limit
+   */
+  _handleFrameSizeError(messageLength) {
+    const errorMsg = `The frame is bigger than expected. Length: ${util.toReadableSizeString(messageLength)}, max: ${util.toReadableSizeString(this.maxFrameLength)}`;
+    log.error(errorMsg);
+    log.error(`Service: ${this.serviceName}, Frame offset: ${this.lengthFieldOffset}, Frame length: ${this.lengthFieldLength}`);
+    
+    // Suggest debugging steps
+    log.error('To debug this issue, set environment variables:');
+    log.error('  APPIUM_IOS_DEBUG_FRAME_SIZE=true');
+    log.error('  APPIUM_IOS_LOG_LARGE_FRAMES=true');
+    log.error('  APPIUM_IOS_MONITOR_MEMORY=true');
+    
+    throw new Error(errorMsg);
   }
 
   _readLength (data, index, littleEndian) {
