+++ lib/utilities.js
@@ -3,24 +3,9 @@ import { upgradeToSSL, enableSSLHandshakeOnly } from './ssl-helper';
 import _ from 'lodash';
 import { log } from './logger';
 import { findDeveloperImage } from './imagemounter/utils/list_developer_image';
+import { util } from '@appium/support';
 
-/**
- * Converts bytes to a human-readable size string
- * @param {number} bytes - Number of bytes to convert
- * @param {number} [decimals=2] - Number of decimal places
- * @returns {string} Human-readable size string (e.g., "1.23 MB", "456.78 KB")
- */
-function toReadableSizeString(bytes, decimals = 2) {
-  if (bytes === 0) return '0 Bytes';
-  
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
-  
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-  
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-}
+
 
 // https://github.com/samdmarshall/iOS-Internals/blob/master/lockbot/lockbot/lockdown_keys.h
 const LOCKDOWN_REQUEST = {
@@ -271,8 +256,24 @@ async function fetchImageFromGithubRepo(udid, opts) {
 }
 
 
+/**
+ * Get maxFrameLength from environment variable or use default value
+ * @param {number} defaultSize - Default frame size in bytes
+ * @returns {number} Configured frame size in bytes
+ */
+function getMaxFrameLength(defaultSize) {
+  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
+  if (envValue) {
+    const parsed = parseInt(envValue, 10);
+    if (!isNaN(parsed) && parsed > 0) {
+      return parsed;
+    }
+  }
+  return defaultSize;
+}
+
 export {
   getConnectedDevices, getOSVersion, getDeviceName, getDeviceTime,
   startLockdownSession, connectPort, connectPortSSL, getDeviceInfo,
-  fetchImageFromGithubRepo, toReadableSizeString
+  fetchImageFromGithubRepo, getMaxFrameLength
 };
