+++ lib/webinspector/index.js
@@ -9,23 +9,13 @@ import { util } from '@appium/support';
 import { MB } from '../constants';
 import { log } from '../logger';
 import { BaseServiceSocket } from '../base-service';
+import { getMaxFrameLength } from '../utilities';
 
 
 const WEB_INSPECTOR_SERVICE_NAME = 'com.apple.webinspector';
 const MAX_FRAME_SIZE = 20 * MB;
 
-// Get maxFrameLength from environment variable or use default
-const getMaxFrameLength = () => {
-  const envValue = process.env.APPIUM_IOS_MAX_FRAME_SIZE;
-  if (envValue) {
-    const parsed = parseInt(envValue, 10);
-    if (!isNaN(parsed) && parsed > 0) {
-      return parsed;
-    }
-    log.warn(`Invalid APPIUM_IOS_MAX_FRAME_SIZE value: ${envValue}, using default: ${MAX_FRAME_SIZE}`);
-  }
-  return MAX_FRAME_SIZE;
-};
+
 
 const PARTIAL_MESSAGE_SUPPORT_DEPRECATION_VERSION = 11;
 
@@ -82,7 +72,7 @@ class WebInspectorService extends BaseServiceSocket {
       verbose = false,
       verboseHexDump = false,
       socketClient,
-      maxFrameLength = getMaxFrameLength(),
+      maxFrameLength = getMaxFrameLength(MAX_FRAME_SIZE),
     } = opts;
 
     super(socketClient);
